dom/media/webaudio/test/test_pannerNodePassThrough.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test PannerNode with passthrough</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="webaudio.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">

var gTest = {
  length: 2048,
  numberOfChannels: 1,
  createGraph: function(context) {
    var source = context.createBufferSource();

    var panner = context.createPanner();

    source.buffer = this.buffer;

    source.connect(panner);

    context.listener.setOrientation(0, 6.311749985202524e+307, 0, 0.1, 1000, 0);
    context.listener.setOrientation(0, 0, -6.311749985202524e+307, 0, 0, 6.311749985202524e+307);
    panner.setPosition(2, 0, 0);
    panner.rolloffFactor = 0;
    panner.panningModel = "equalpower";

    var pannerWrapped = SpecialPowers.wrap(panner);
    ok("passThrough" in pannerWrapped, "PannerNode should support the passThrough API");
    pannerWrapped.passThrough = true;

    source.start(0);
    return panner;
  },
  createExpectedBuffers: function(context) {
    this.buffer = context.createBuffer(1, 2048, context.sampleRate);
    for (var i = 0; i < 2048; ++i) {
      this.buffer.getChannelData(0)[i] = Math.sin(440 * 2 * Math.PI * i / context.sampleRate);
    }

    return [this.buffer];
  },
};

runTest();

</script>
</pre>
</body>
</html>
